+++ src/files/BrsFile.spec.ts
@@ -203,6 +203,19 @@ describe('BrsFile', () => {
             expect(results).to.be.empty;
         });
 
+        it('does provide intellisence for labels only after a goto keyword', () => {
+            //eslint-disable-next-line @typescript-eslint/no-floating-promises
+            program.addOrReplaceFile({ src: `${rootDir}/source/main.brs`, dest: 'source/main.brs' }, `
+                sub Main(name as string)
+                    something:
+                    goto \nend sub
+            `);
+
+            let results = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(3, 25));
+            expect(results.length).to.equal(1);
+            expect(results[0]?.label).to.equal('something');
+        });
+
     });
 
     describe('comment flags', () => {
@@ -210,7 +223,7 @@ describe('BrsFile', () => {
             it('disables critical diagnostic issues', () => {
                 program.addOrReplaceFile('source/main.brs', `
                     sub main()
-                        Dim requestData[requestList.count()]
+                        Dim requestData
                     end sub
                 `);
                 //should have an error
@@ -220,7 +233,7 @@ describe('BrsFile', () => {
                 program.addOrReplaceFile('source/main.brs', `
                     sub main()
                         'bs:disable-next-line
-                        Dim requestData[requestList.count()]
+                        Dim requestData
                     end sub
                 `);
                 //should have an error
@@ -1729,6 +1742,31 @@ describe('BrsFile', () => {
                 sub NameA_NameB_alert()
                 end sub
             `, 'trim', 'source/main.bs');
+        });
+
+        it('transpiles dim', () => {
+            testTranspile(`Dim c[5]`, `dim c[5]`);
+            testTranspile(`Dim c[5, 4]`, `dim c[5, 4]`);
+            testTranspile(`Dim c[5, 4, 6]`, `dim c[5, 4, 6]`);
+            testTranspile(`Dim requestData[requestList.count()]`, `dim requestData[requestList.count()]`);
+            testTranspile(`Dim requestData[1, requestList.count()]`, `dim requestData[1, requestList.count()]`);
+            testTranspile(`Dim requestData[1, requestList.count(), 2]`, `dim requestData[1, requestList.count(), 2]`);
+            testTranspile(`Dim requestData[requestList[2]]`, `dim requestData[requestList[2]]`);
+            testTranspile(`Dim requestData[1, requestList[2]]`, `dim requestData[1, requestList[2]]`);
+            testTranspile(`Dim requestData[1, requestList[2], 2]`, `dim requestData[1, requestList[2], 2]`);
+            testTranspile(`Dim requestData[requestList["2"]]`, `dim requestData[requestList["2"]]`);
+            testTranspile(`Dim requestData[1, requestList["2"]]`, `dim requestData[1, requestList["2"]]`);
+            testTranspile(`Dim requestData[1, requestList["2"], 2]`, `dim requestData[1, requestList["2"], 2]`);
+            testTranspile(`Dim requestData[1, getValue(), 2]`, `dim requestData[1, getValue(), 2]`);
+            testTranspile(`
+                Dim requestData[1, getValue({
+                    key: "value"
+                }), 2]
+            `, `
+                dim requestData[1, getValue({
+                    key: "value"
+                }), 2]
+            `);
         });
 
         it('transpiles calls to fully-qualified namespaced functions', () => {