+++ CHANGELOG.md
@@ -6,6 +6,56 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [0.30.8] - 2021-02-12
+### Changed
+ - add additional logging in `Program.removeFile`
+### Fixed
+ - Fix watcher bug on windows devices
+ - Don't mangle xml scripts during transpile
+
+
+
+## [0.30.7] - 2021-02-11
+### Fixed
+ - bug in `getSignatureHelp` that wouldn't work for function calls with no leading whitespace. ([#307](https://github.com/rokucommunity/brighterscript/issues/307))
+
+
+
+## [0.30.6] - 2021-02-07
+### Fixed
+ - bad transpile for nested class method super calls
+ - SceneGraph node attributes being wrongly removed when modifying attributes
+
+
+
+## [0.30.5] - 2021-02-03
+### Added
+ - syntax support for dim statements
+ - completion and code navigation for labels
+### Fixed
+ - exception related to signature help when writing comments
+
+
+
+## [0.30.4] - 2021-02-02
+### Fixed
+ - fixed crash during validation caused by a missing function body when parsing malformed code
+
+
+
+## [0.30.3] - 2021-02-01
+### Fixed
+ - performance issue when transpiling larger projects, even when no brighterscript code was used
+
+
+
+## [0.30.2] - 2021-01-31
+### Fixed
+ - xml parse error crashing validation ((#294)[https://github.com/rokucommunity/brighterscript/pull/294])
+ - better handling for `createStringLiteral` ((#292)[https://github.com/rokucommunity/brighterscript/pull/292])
+
+
+
 ## [0.30.1] - 2021-01-29
 ### Fixed
  - bug that would crash while transpiling if a script tag didn't have a `type` attribute
@@ -880,3 +930,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.29.0]:   https://github.com/rokucommunity/brighterscript/compare/v0.28.2...v0.29.0
 [0.30.0]:   https://github.com/rokucommunity/brighterscript/compare/v0.29.0...v0.30.0
 [0.30.1]:   https://github.com/rokucommunity/brighterscript/compare/v0.30.0...v0.30.1
+[0.30.2]:   https://github.com/rokucommunity/brighterscript/compare/v0.30.1...v0.30.2
+[0.30.3]:   https://github.com/rokucommunity/brighterscript/compare/v0.30.2...v0.30.3
+[0.30.4]:   https://github.com/rokucommunity/brighterscript/compare/v0.30.3...v0.30.4
+[0.30.5]:   https://github.com/rokucommunity/brighterscript/compare/v0.30.4...v0.30.5
+[0.30.6]:   https://github.com/rokucommunity/brighterscript/compare/v0.30.5...v0.30.6
+[0.30.7]:   https://github.com/rokucommunity/brighterscript/compare/v0.30.6...v0.30.7
+[0.30.8]:   https://github.com/rokucommunity/brighterscript/compare/v0.30.7...v0.30.8